Modularized algorithm
