+++ server/engines/similars.js
@@ -1,28 +1,8 @@
-var _ = require('underscore');
-var async = require('async');
+var redis = require('redis');
+var Promise = require("bluebird");
+Promise.promisifyAll(require("redis"));
 
-/*
-var Similars = function(engine){  // Pass in engine?
-    // Initialize with Engine and db connection    
-        // this.engine = engine;
-        // this.db = ...
-};
-
-Similars.prototype.byUser = function(){
-    // This funciton returns List of Users that is similar to input User.  Retrieve from database
-
-};
-
-Similars.prototype.update = function(){
-    // Async AUTO
-        //Arg1: Obj with:   // [BRYAN NOTE: ] 
-            // userLike:  function(done) => this.engine.like.itemByUser(user,done)
-            // userDislike: function(done) => this.engine.dislike.itemByUser(user,done)
-
-
-};
-*/
-var Similars = function(e, db) {
+var Similars = function(db) {
   this.db = db;
   this.similars = ""; 
 };
@@ -35,6 +15,10 @@ Similars.prototype.byUser = function(userID) {
 Similars.prototype.update = function(userID) {
   var userLikes = userID + ":Likes";
   var userDislikes = userID + ":Dislikes";
+
+  var otherUserList = [];
+  var otherUserScore = [];
+
 
   this.db.sunionstore("userRated", userLikes, userDislikes);
   var that = this;
@@ -50,10 +34,7 @@ Similars.prototype.update = function(userID) {
       that.db.sunionstore("comparisonMembers", "comparisonMembers", restaurantArray[i] + ":Dislikes");
     }
 //    that.db.srem("comparisonMembers", userID);
-    for (i = 0; i < 10000000; i++) {
-        j = 1;
-
-    }
+  
     that.db.smembers("comparisonMembers", function(err, compMembersArray) {
       var comparisonIndex;
       var commonLikes;
@@ -63,10 +44,21 @@ Similars.prototype.update = function(userID) {
       var otherUserLikes;
       var otherUserDislikes;
 
+      var commonLikesArr = [];
+      var commonDislikesArr = [];
+      var conflicts1Arr = [];
+      var conflicts2Arr = [];
+      var allRatedRestaurantsArr = [];
+
+
       for (i = 0; i < compMembersArray.length; i++) {
-        // console.log(compMembersArray.length);
+
+        console.log(compMembersArray.length);
         otherUserLikes = compMembersArray[i] + ":Likes";
-        otherUserDislikes = compMembersArray[i] + ":Dislikes";        
+        otherUserDislikes = compMembersArray[i] + ":Dislikes";
+
+        otherUserList.push(compMembersArray[i]); 
+        console.log(otherUserList);       
         //these are temporary lists, need to clear them somehow
   
         that.db.sinterstore("commonLikes", userLikes, otherUserLikes);
@@ -77,40 +69,45 @@ Similars.prototype.update = function(userID) {
                         userDislikes, otherUserDislikes);
 
         that.db.scard("commonLikes", function(err, commonLikesCount) {
-            console.log("COMMON LIKES:  " + userID + "-  " + commonLikesCount);
-          that.db.scard("commonDislikes", function(err, commonDislikesCount) {
-            that.db.scard("conflicts1", function(err, conflicts1Count) {
-              that.db.scard("conflicts2", function(err, conflicts2Count) {
-                that.db.scard("allRatedRestaurants", function(err, allRatedRestaurantsCount) {
-                  // console.log("HELLO");
-                  // console.log(allRatedRestaurantsCount);
-                  // console.log((Number(commonLikesCount) + Number(commonDislikesCount) -
-                  //              Number(conflicts1Count) - Number(conflicts2Count)) / Number(allRatedRestaurantsCount));
+          commonLikesArr.push(commonLikesCount);
+        });
 
-                });
+        that.db.scard("commonDislikes", function(err, commonDislikesCount) {
+          commonDislikesArr.push(commonDislikesCount);
+        });
+
+        that.db.scard("conflicts1", function(err, conflicts1Count) {
+          conflicts1Arr.push(conflicts1Count);
+        });
+
+        that.db.scard("conflicts2", function(err, conflicts2Count) {
+          conflicts2Arr.push(conflicts2Count);
+        });
+
+        that.db.scard("allRatedRestaurants", function(err, allRatedRestaurantsCount) {
+          allRatedRestaurantsArr.push(allRatedRestaurantsCount);
+          if (compMembersArray.length === commonLikesArr.length) {
+            for (var k = 0; k < commonLikesArr.length; k++) {
+              // console.log(" commonLikesCount:  " + commonLikesArr[k] +
+              //     " commonDislikes:  " + commonDislikesArr[k] +
+              //     " conflicts1:  " + conflicts1Arr[k] +
+              //     " conflicts2:  " + conflicts2Arr[k] +
+              //     " allRatedRestaurants:  " + allRatedRestaurantsArr[k]);
+              comparisonIndex = (Number(commonLikesArr[k]) + Number(commonDislikesArr[k]) -
+                       Number(conflicts1Arr[k]) - Number(conflicts2Arr[k])) / Number(allRatedRestaurantsArr[k]);
+              console.log("COMPARISON INDEX" + userID + ":  " + compMembersArray[k] + "  " + comparisonIndex);
+              that.db.zadd(userID + ":Similars", comparisonIndex, compMembersArray[k]);
+              that.db.zrange(userID + ":Similars", 0, -1, function(err, answer) {
+                // console.log ("DO WE GET HERE");
+                // console.log(answer);
               });
-            });
-          });
+            }
+          }
         });
       }
     });
-  });
+});
 };
-
 
 module.exports = Similars;
 
-
-/*  
-ASYNC auto;
-async.auto({
-  readData: async.apply(fs.readFile, 'data.txt', 'utf-8')
-}, callback);
-
-ASYNC Map:
-
-async.map(['file1','file2','file3'], fs.stat, function(err, results){
-    // results is now an array of stats for each file 
-});
-
-*/
